+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changes
 * [UI/Developer]: Updated `fstream` to fix vulnerability.
 * [Developer]: Update pom file to reflect active IRIDA developers.
 * [UI]: Fixed bug where a new metadata template could not be created. (19.05.1)
+* [UI/Developer]: Updated to latest `react-redux` to use new hooks API.
 * [UI]: Added the ability to delete all the data in a column in the line list table.
 
 19.01 to 19.05